Move TabsContainerItemProps to dedicated type file
