+++ apps/src/shared/gamma/containers/tabs/TabsContainerItem.types.ts
@@ -1,0 +1,9 @@
+import type { TabRoute, TabsNavigationMethods } from './TabsContainer.types';
+
+export type TabsContainerItemProps = {
+  route: TabRoute;
+  navMethods: TabsNavigationMethods;
+  isSelected: boolean;
+  pendingForUpdate: boolean;
+}
+
